# Skip in-compressible data (32x) faster