@pierrec 2019-07-01
Again thanks Klaus. I used to have something like this but it failed to compress some patterns. Your implementation seems to work on those.