@klauspost 2019-07-01
@pierrec - Ah - I remember seeing this in your library first, and thought I was mis-remembering when it wasn't there.

I assume it came from the C library. In general, the gains are so big and the downsides so unlikely that it is well worth having in there. This might also be more conservative than the initial implementation.